style(download-button): removed heading for download button, moved download_csv()
